+++ Changelog.md
@@ -6,7 +6,7 @@ Expect active development and potentially significant breaking changes in the `0
 
 ### v0.5.1
 
-- Feature: Added link to [recompose](https://github.com/acdlite/recompose) to use the `compose` function. This makes it easy to combine multiple queries on a single component.
+- Feature: Added link to [recompose](https://github.com/acdlite/recompose) to use the `compose` function. This makes it easy to combine multiple queries on a single component. [#194](https://github.com/apollostack/react-apollo/pull/194)
 
 ### v0.5.0